+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.13.10] - 2024-FEBRUARY-9
+
+- Remove unused internal APIs
+- Add Fraxtal testnet
+
 ## [0.13.8] - 2023-DECEMBER-18
 
 - Add `Immutable` to mainnet configs; update `Immutable` testnet RPC